Added First Message support to Twitch user messages.
Twitch now sends to moderators and stream owners a new tag with all chat messages, "first-msg=N" where N is 0 or 1. If it's 1, this is the first time this user chatted on the target user's stream. This CL adds support for parsing that tag, and adds a new boolean to the private message struct called FirstMessage to reflect this status.